Reorganized old admin management components
