+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/AdminsAndModerators/ManageRoles/ManageRoles.scss
@@ -1,4 +1,4 @@
-@import '../../shared';
+@import '../../../../../../styles/shared';
 
 .ManageRoles {
   display: flex;
