+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/AdminsAndModerators/ManageRoles/index.tsx
@@ -1,0 +1,3 @@
+import { ManageRoles } from './ManageRoles';
+
+export default ManageRoles;
