+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/AdminsAndModerators/UpgradeRolesForm/UpgradeRolesForm.scss
@@ -1,4 +1,4 @@
-@import '../../shared';
+@import '../../../../../../styles/shared';
 
 .UpgradeRolesForm {
   display: flex;
