+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/AdminsAndModerators/UpgradeRolesForm/index.tsx
@@ -1,0 +1,3 @@
+import { UpgradeRolesForm } from './UpgradeRolesForm';
+
+export default UpgradeRolesForm;
