+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Integrations/Discord/CallbackPage.tsx
@@ -2,10 +2,10 @@ import axios from 'axios';
 import { useCommonNavigate } from 'navigation/helpers';
 import React, { useCallback, useEffect, useState } from 'react';
 import app from 'state';
-import { PageNotFound } from '../404';
-import { PageLoading } from '../loading';
+import { PageNotFound } from '../../../404';
+import { PageLoading } from '../../../loading';
 
-const DiscordCallbackPage = () => {
+const CallbackPage = () => {
   const navigate = useCommonNavigate();
   const [failed, setFailed] = useState(false);
   const [failureMessage, setFailureMessage] = useState<string>('');
@@ -68,4 +68,5 @@ const DiscordCallbackPage = () => {
     <PageLoading message="Connecting Discord" />
   );
 };
-export default DiscordCallbackPage;
+
+export default CallbackPage;
