+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Integrations/Discord/Discord.scss
@@ -20,7 +20,7 @@
     }
 
     .channels {
-      .DiscordForumConnections {
+      .DiscordConnections {
         display: flex;
         flex-direction: column;
         gap: 8px;
